pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claim-extractor"
version = "0.1.0"
description = "Extract structured claims from text and PDFs"
requires-python = ">=3.8"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "openai>=1.0.0",
    "requests>=2.25.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.0.0",
    "pypdf>=3.0.0",
    "pdfminer.six>=20221105",
]

[tool.hatch.build.targets.wheel]
packages = [
    "src/claim_extractor",
    "src/pdf_parser",
]

[tool.pytest.ini_options]
markers = [
    "openai: tests that use OpenAI API",
    "anthropic: tests that use Anthropic API",
    "integration: tests that use any real API",
    "slow: marks tests as slow",
]
